Fold UvSO2TensorProduct into tece/interaction.py.
Drop the leftover fused.py split now that only one message class remains.
