+++ docs/user_guide/models/architecture_catalog.rst
@@ -87,7 +87,7 @@ External models are declared under :code:`Modelhub.external_FFs` with the same :
 
 **TACE** (:code:`architecture: tace`) is a native port of Xu et al. (arXiv:2509.14961; Cartesian-3j arXiv:2512.16882) under :code:`enerzyme/models/tace/` (:code:`core.py`, :code:`interaction.py`), adapted from `xvzemin/tace <https://github.com/xvzemin/tace>`_ (MIT). Shared flat-Irreps helpers live in :code:`enerzyme/models/e3nn_nn/`; radial MLPs use :code:`blocks.radial_mlp.RadialMLP`. The Core is the ACE message/product stack only (edge update + CgtpInteraction / Cartesian ICT + CgtpACE); embeddings, property heads, ZBL/LES/UIE stay outside. Switch backends with Core param :code:`tensor_basis: spherical|cartesian` (default spherical). TECE / SO2 / RRA live in the separate :code:`tece` architecture. Emits :code:`atom_feature` with :code:`feature_irreps` / :code:`dim_feature_out` for :code:`SimpleReadout`. Examples: :code:`enerzyme/config/tace_layers_example.yaml`, :code:`tace_cartesian_layers_example.yaml`. Tests: :code:`test/test_tace_core.py`, :code:`test/test_tace_spherical_ops.py`, :code:`test/test_tace_cartesian_ops.py`. Enerzymette only needs :code:`architecture: tace` plus a resolved :code:`config.yaml`.
 
-**TECE** (:code:`architecture: tece`) is a native port of Xu et al. (arXiv:2607.10664) under :code:`enerzyme/models/tece/` (:code:`core.py`, :code:`interaction.py`, :code:`fused.py`), adapted from the same `xvzemin/tace <https://github.com/xvzemin/tace>`_ v0.2.0 SO2 path (MIT). The Core runs a tensor seed (scalar → equivariant) then uvSO2 layers with **Edge Cluster Expansion** (:code:`ComplexProductBasis`) and **Radial Rotary Attention** (complex Q·K + radial phase), followed by node-side :code:`CgtpACE` reused from TACE. Shared SO(2) primitives (:code:`WignerD`, :code:`uvSO2Linear`, :code:`SO2Gate`, :code:`LayoutTransform`) live in :code:`enerzyme/models/so3/`. Requires :code:`Lmax == lmax`. Distinct from EFA's Euclidean RoPE. Emits :code:`atom_feature` for :code:`SimpleReadout`. Example: :code:`enerzyme/config/tece_layers_example.yaml`. Tests: :code:`test/test_tece_core.py`, :code:`test/test_tece_ops.py`. Enerzymette only needs :code:`architecture: tece` plus a resolved :code:`config.yaml`.
+**TECE** (:code:`architecture: tece`) is a native port of Xu et al. (arXiv:2607.10664) under :code:`enerzyme/models/tece/` (:code:`core.py`, :code:`interaction.py`), adapted from the same `xvzemin/tace <https://github.com/xvzemin/tace>`_ v0.2.0 SO2 path (MIT). The Core runs a tensor seed (scalar → equivariant) then uvSO2 layers with **Edge Cluster Expansion** (:code:`ComplexProductBasis`) and **Radial Rotary Attention** (complex Q·K + radial phase), followed by node-side :code:`CgtpACE` reused from TACE. Shared SO(2) primitives (:code:`WignerD`, :code:`uvSO2Linear`, :code:`SO2Gate`, :code:`LayoutTransform`) live in :code:`enerzyme/models/so3/`. Requires :code:`Lmax == lmax`. Distinct from EFA's Euclidean RoPE. Emits :code:`atom_feature` for :code:`SimpleReadout`. Example: :code:`enerzyme/config/tece_layers_example.yaml`. Tests: :code:`test/test_tece_core.py`, :code:`test/test_tece_ops.py`. Enerzymette only needs :code:`architecture: tece` plus a resolved :code:`config.yaml`.
 
 **So3krates** (:code:`architecture: so3krates`) is a native port of Frank et al. (NeurIPS 2022) under :code:`enerzyme/models/so3krates/`, following the So3krates-torch EuclideanTransformer (fused FeatureBlock + GeometricBlock + InteractionBlock). Shared :code:`RealSphericalHarmonics` and :code:`L0Contraction` live in :code:`enerzyme/models/so3/`. The Core emits :code:`atom_feature` (invariant stream ``x``, :code:`feature_irreps: "Fx0e"`) and :code:`atom_sphere_feature` (SPHC ``χ`` with shape ``[N, m_tot]``, **not** eSCN/EquiformerV2's ``[N, (lmax+1)^2, C]`` — :code:`SphereSampleReadout` does not apply). Default stacks use :code:`BernsteinRBF` + :code:`SimpleReadout` + :code:`EnergyReduce` + :code:`Force`. Long-range physics (ZBL / electrostatics / dispersion) belong in post-core layers, not inside the Core. Example: :code:`enerzyme/config/so3krates_layers_example.yaml`. Parity: :code:`test/test_so3krates_parity_ops.py`.
